Move EditorStore to src/contexts/Store
