+++ .vscode/settings.json
@@ -1,5 +1,12 @@
 {
+  "editor.formatOnSave": true,
+  "editor.fontSize": 15,
+  "editor.fontFamily": "Noto Sans Mono, monospace",
   "editor.tabSize": 2,
+  "editor.minimap.enabled": false,
+  "workbench.editor.openSideBySideDirection": "right",
+  "workbench.colorTheme": "CodeSandbox Black",
+  "editor.renderWhitespace": "all",
   "editor.rulers": [
     80,
     100
