+++ package.json
@@ -13,12 +13,12 @@
     "react": "16.12.0",
     "react-dom": "16.12.0",
     "react-scripts": "3.3.0",
-    "styled-components": "4.4.1"
+    "styled-components": "4.4.1",
+    "typescript": "3.8.0-dev.20191121"
   },
   "devDependencies": {
     "@types/react": "16.8.8",
-    "@types/react-dom": "16.8.2",
-    "typescript": "3.3.3"
+    "@types/react-dom": "16.8.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -26,11 +26,6 @@
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ],
+  "browserslist": [">1%"],
   "license": "MIT"
 }
