+++ src/components/Editor/index.ts
@@ -1,4 +1,3 @@
 export { Editor } from "./Editor";
 export { GlobalStyle } from "./GlobalStyle";
 export { Wrapper } from "./Wrapper";
-export { useStore, Context, Provider } from "./Store";
