+++ src/components/Toolbar/Toolbar.tsx
@@ -1,46 +1,9 @@
 import React from "react";
-import Draft from "draft-js";
 import { Icon } from "../Icon";
 import { ToolbarButton } from "./ToolbarButton";
 import { Wrapper } from "./Wrapper";
-import { useStore } from "../Editor";
-import { replaceDraftState } from "../Editor/Store";
-
-enum InlineStyles {
-  Bold = "BOLD",
-  Italic = "ITALIC",
-  Underline = "UNDERLINE",
-  StrikeThrough = "STRIKETHROUGH",
-}
-
-interface ButtonOption {
-  style: InlineStyles;
-  label: string;
-  icon: string;
-}
-
-const BUTTON_OPTIONS: ButtonOption[] = [
-  {
-    style: InlineStyles.Bold,
-    label: "Bold",
-    icon: "format_bold",
-  },
-  {
-    style: InlineStyles.Italic,
-    label: "Italic",
-    icon: "format_italic",
-  },
-  {
-    style: InlineStyles.Underline,
-    label: "Underline",
-    icon: "format_underline",
-  },
-  {
-    style: InlineStyles.StrikeThrough,
-    label: "Strike-through",
-    icon: "format_strikethrough",
-  },
-];
+import { useStore, toggleInline } from "../../contexts/Store";
+import { INLINE_OPTIONS } from "../Editor/Inlines";
 
 export const Toolbar = () => {
   const {
@@ -50,18 +13,12 @@ export const Toolbar = () => {
 
   return (
     <Wrapper>
-      {BUTTON_OPTIONS.map(({ label, style, icon }) => (
+      {INLINE_OPTIONS.map(({ label, style, icon }) => (
         <ToolbarButton
           key={style}
           title={label}
           active={draftState.getCurrentInlineStyle().has(style)}
-          onClick={() =>
-            dispatch(
-              replaceDraftState(
-                Draft.RichUtils.toggleInlineStyle(draftState, style),
-              ),
-            )
-          }
+          onClick={() => dispatch(toggleInline(style))}
         >
           <Icon>{icon}</Icon>
         </ToolbarButton>
