+++ src/contexts/Store/index.ts
@@ -1,0 +1,2 @@
+export { Context, useStore, Provider } from "./Context";
+export { toggleInline, replaceDraftState, keyCommand } from "./actions";
