+++ src/types.ts
@@ -1,0 +1,6 @@
+export enum InlineStyles {
+  Bold = "BOLD",
+  Italic = "ITALIC",
+  Underline = "UNDERLINE",
+  StrikeThrough = "STRIKETHROUGH",
+}
