+++ tsconfig.json
@@ -5,7 +5,7 @@
     "compilerOptions": {
         "lib": [
             "dom",
-            "es2015"
+            "es2017"
         ],
         "jsx": "react",
         "strict": true,
